change structure graph class

+++ main.h
@@ -7,7 +7,7 @@
 int n_nodes = 10;
 int n_drones = 1;
 int n_batteries = n_drones;
-int budget = 1;
+double budget = 1;
 std::string graph_file;
 
 #endif // MAIN_H
